socket: decrypt client initial packets for user defined selectors
reuseport socket option allows users to select from multiple sockets
that listen to the same address:port based on the client initial
packet in QUIC.

However when calling reuseport_select_sock() in quic_sock_lookup(),
the client initial packet is not yet decrypted, which causes that
users have to do decryption before accessing the data like ALPN
in the TLS message carried in the crypto frame.

The decryption could be complex for users due to the key deriving
and both header skcipher and payload aead decryption, which is
relative easy to be done in QUIC stack with these APIs in crypto.c.

This patch adds quic_packet_early_decrypt() to do this decryption
before calling reuseport_select_sock() in quic_sock_lookup() to
provide user defined selectors with a plain QUIC packet.

Signed-off-by: Xin Long <lucien.xin@gmail.com>