tests: return 0 when no ticket is received in http_client_handshake
In interop tests, some implementation does not send ticket in the 2nd
connection, and we must not return err for this.

Also, the handshake error is returned by http_client_handshake() not
in the errno, so in http09_client() and http3_client(), get the err
from the return value instead of the errno.

Signed-off-by: Xin Long <lucien.xin@gmail.com>